@@ -1,284 +1,32 @@
 """2ch.hk API"""
 
-__all__ = ('DvachApi', 'Board', 'Thread', 'Post', 'Message', 'BOARDS', 'BOARDS_ALL')
+__all__ = ('DvachApi', 'Message', 'CHAN_URL')
 
-from posixpath import join as url_join
-
-import requests
-from addict import Dict
-from simplejson import JSONDecodeError
-
-from .utils import listmerge
-
-BOARDS = {
-    'thematics': ['bi', 'biz', 'bo', 'c', 'em', 'fa', 'fiz', 'fl', 'ftb', 'hi', 'me', 'mg', 'mlp', 'mo', 'ne', 'psy',
-                  're', 'sf', 'sci', 'sn', 'sp', 'spc', 'tv', 'un', 'w', 'wh', 'wm', 'mov', 'rf', 'mu', 'au', 'zog',
-                  'o'],
-    'creation': ['di', 'de', 'diy', 'mus', 'pa', 'p', 'wp', 'wrk'],
-    'tech': ['hw', 'pr', 'ra', 's', 't', 'gd', 'mobi'],
-    'politics': ['po', 'news'],
-    'games': ['bg', 'cg', 'mmo', 'tes', 'vg', 'wr', 'moba', 'v', 'pok', 'ruvn'],
-    'japanese': ['a', 'fd', 'ja', 'ma', 'vn'],
-    'other': ['b', 'd', 'soc', 'r', 'abu', 'media'],
-    'adults': ['fag', 'fg', 'fur', 'gg', 'ga', 'h', 'ho', 'sex', 'fet', 'e', 'hc', 'guro', 'vape'],
-    'user': ['ew', 'hh', 'pvc', 'ph', 'tr', 'dom', 'izd', 'td', 'mc', 'aa', 'rm', 'to', 'web', 'br', 'trv', 'gb', 'fs',
-             'cul', 'out', 'old', 'cc', 'ussr', 'jsf', 'ukr', 'sw', 'law', 'm', 'ya', 'r34', 'qtr4', 'wow', 'gabe',
-             'cute', 'by', 'se', 'kz', '8', 'es', 'alco', 'brg', 'mlpr', 'ro', 'who', 'srv', 'asmr', 'dr', 'electrach',
-             'ing', 'got', 'crypt', 'socionics', 'lap', 'smo', 'hg', 'sad', 'fi', 'nvr', 'ind', 'ld', 'fem', 'gsg',
-             'kpop', 'vr', 'arg', 'char', 'obr', 'hv', '2d', 'wwe', 'ch', 'int', 'math', 'test']
-}
-
-BOARDS_ALL = listmerge(BOARDS)
-
-URL = 'https://2ch.hk/'
-
-
-class ApiSession:
-    HEADERS = {
-        'User-agent': 'Mozilla/5.0 (Windows NT 6.1; rv:52.0) '
-                      'Gecko/20100101 Firefox/52.0'
-    }
-
-    def __init__(self, proxies=None, headers=None):
-        self.session = requests.Session()
-        self.session.headers.update(headers if headers else self.HEADERS)
-        if proxies:
-            self.session.proxies.update(proxies)
-
-    def get(self, *args):
-        """
-        Get url
-        :param args: args for request
-        :return:
-        """
-        url = url_join(URL, *args)
-        try:
-            response = self.session.get(url=url)
-        except Exception as e:
-            print('Something goes wrong:', e)
-            return None
-        else:
-            try:
-                return Dict(response.json())
-            except JSONDecodeError:
-                return response
-
-    def post(self, **kwargs):
-        """
-        Post url
-        :param kwargs: kwargs for request
-        :return: request response
-        """
-        url = url_join(URL, kwargs['url'])
-        try:
-            response = self.session.post(url=url, data=kwargs['data'], files=kwargs['files'])
-        except Exception as e:
-            print('Something goes wrong:', e)
-            return None
-        else:
-            try:
-                return Dict(response.json())
-            except JSONDecodeError:
-                return response
-
-    def update_headers(self, headers):
-        if headers:
-            self.session.headers.clear()
-            self.session.headers.update(headers if headers else self.HEADERS)
-
-    def update_proxies(self, proxies):
-        if proxies:
-            self.session.proxies.clear()
-            self.session.proxies.update(proxies)
-
-
-class Board:
-    """Board object"""
-
-    __rows__ = ('bump_limit', 'category', 'default_name', 'enable_dices',
-                'enable_flags', 'enable_icons', 'enable_likes',
-                'enable_names', 'enable_oekaki', 'enable_posting',
-                'enable_sage', 'enable_shield', 'enable_subject', 'enable_thread_tags',
-                'enable_trips', 'icons', 'id', 'name', 'pages', 'sage', 'tripcodes')
-
-    def __init__(self, settings):
-        """
-        Create object from dict with settings
-        :param settings: dict with settings
-        """
-        self.id = None
-        for key, value in settings.items():
-            setattr(self, key, value)
-
-    def __repr__(self):  # pragma: no cover
-        return '<Settings: {board}>'.format(board=self.id)
-
-
-class Thread:
-    """Thread object"""
-
-    def __init__(self, thread):
-        """
-        Create object from dict with thread info
-        :param thread: dict with thread info
-        """
-        self.reply_count = int(thread['posts_count'])
-        self.post = Post(thread)
-        self.num = self.post.num
-
-    def __repr__(self):
-        return '<Thread: {num}>'.format(num=self.num)
-
-
-class Post:
-    """Post object"""
-
-    __rows__ = ('banned', 'closed', 'comment', 'date', 'email',
-                'endless', 'files', 'lasthit', 'name', 'num',
-                'number', 'op', 'parent', 'sticky', 'subject',
-                'tags', 'timestamp', 'trip')
-
-    def __init__(self, post):
-        """
-        Create object from dict with post info
-        :param post: dict with post info
-        """
-        self.comment = None
-        self.num = None
-        for key, value in post.items():
-            setattr(self, key, value)
-
-    def __repr__(self):
-        return '<Post: {num}>'.format(num=self.num)
-
-
-class Message:
-    """Message object"""
-
-    # формирование пайлоада сообщения
-    def __init__(self, board_id, thread_id, comment='', email='', subject='', name='', sage=False, files=None):
-        """
-        Офрмируется пайлоад, добавляются файлы при наличии
-        :param board_id: Доска
-        :param thread_id: Номер треда
-        :param comment: Тело сообщения
-        :param email: email
-        :param subject: Тема
-        :param name: Имя
-        :param sage: Вкл/Выкл сажу
-        :param files: Список файлов 1-8 (обрежется до 1-4 без пасскода)
-        """
-        # формируем пайлоад
-        self.payload = Dict({
-            'json': 1,
-            'task': 'post',
-            'board': board_id,
-            'thread': thread_id,
-            'email': email,
-            'name': name,
-            'subject': subject,
-            'comment': comment,
-            'sage': 1 if sage else 0
-        })
-        self.files = {}
-        # Добавляем файл при наличии
-        # TODO: Однострочник не взлетел
-        if files and 0 < len(files) <= 8:
-            try:
-                for file_name in files:
-                    file = open(file_name, 'rb')
-                    self.files[
-                        file.name] = file.read()  # Можно передавать любой filename, не обязательно imageX. Tested[+]
-                    file.close()  # читаем-закрываем сразу
-            except Exception as e:
-                print("IO error:", e)
-        else:
-            self.files = {'': ''}
-
-    def __repr__(self):
-        return f'<Message: {self.payload}, Files: {self.files.keys() if self.files else 0}>'
-
-
-class Captcha:
-    """Объект Captcha"""
-
-    def __init__(self, captcha_id):
-        self.captcha_type = '2chaptcha'
-        self.captcha_id = captcha_id
-        self.captcha_value = None
-
-    def set_answer(self, answer):
-        self.captcha_value = int(answer)
-
-    def __repr__(self):
-        return f'Captcha type: {self.captcha_type}, Captcha ID: {self.captcha_id}, Captcha value: {self.captcha_value}>'
-
-
-class CaptchaHelper:
-    """Класс помошник при работе с капчёй."""
-
-    def __init__(self, session):
-        """
-        Инициализирует сессию для капчи
-        :param session: Сессия ApiSession
-        """
-        self.__Session = session
-
-    def get_captcha(self):
-        """
-        Метод отвечает за получение капчи
-        :return: Объект типа Captcha
-        """
-        captcha_response = self.__Session.get(f'api/captcha/2chaptcha/service_id')
-        if captcha_response.result == 1:
-            captcha_id = captcha_response.id
-
-            return Captcha(captcha_id)
-        else:
-            return False
-
-    def get_captcha_img(self, captcha):
-        """
-        Метод для получения изображение капчи
-        :param captcha: Объект типа Captcha
-        :return: Словарь с ссылкой на капчу и её бинарное представление
-        """
-        captcha_image = self.__Session.get(f'api/captcha/2chaptcha/image/{captcha.captcha_id}').content
-        url = f'{URL}api/captcha/2chaptcha/image/{captcha.captcha_id}'
-
-        return Dict({'url': url, 'binary': captcha_image})
-
-    def check_captcha(self, captcha):
-        """
-        Метод отвечает за проверку правельности решения капчи
-        :return: Возвращает True/False в зависимости от праильности решения капчи
-        """
-        response = self.__Session.get(f'api/captcha/2chaptcha/check/{captcha.captcha_id}?value={captcha.captcha_value}')
-
-        if response.result == 1:
-            return True
-        else:
-            return False
+import api2ch.exceptions as ex
+from .captcha import CaptchaHelper
+from .helpers import get_all_board_settings
+from .models import Message, Thread, Post
+from .client import ApiClient
+from .settings import CHAN_URL
 
 
 class DvachApi:
     """Объект DvachApi"""
 
-    _boards = {}
-
-    def __init__(self, board='b', proxies=None, headers=None):
+    def __init__(self, board='b', passcode=None, proxies=None, headers=None):
         """
         Инициализация Api
         :param board: ИД доски
         """
-
-        self.__Session = ApiSession(proxies=proxies, headers=headers)
-        self.__get_all_settings()
+        self.__ApiClient = ApiClient(proxies=proxies, headers=headers)
+        self.boards = get_all_board_settings(self.__ApiClient)
         self.__board = None
         self.board = board
-        self.settings = None
-        self.CaptchaHelper = CaptchaHelper(self.__Session)
+        self.CaptchaHelper = CaptchaHelper(self.__ApiClient)
         self.passcode_data = None
+
+        if passcode:
+            self.auth_passcode(passcode)
 
     @property
     def board(self):
@@ -286,23 +34,10 @@ class DvachApi:
 
     @board.setter
     def board(self, board):
-        if board in self._boards.keys():
-            self.__board = self._boards[board]
+        if board in self.boards.keys():
+            self.__board = self.boards[board]
         else:
             self.__board = None
-
-    def __get_all_settings(self):
-        all_settings = self.__Session.get('makaba/mobile.fcgi?task=get_boards')
-
-        for key in all_settings.keys():
-            for settings in all_settings[key]:
-                self._boards[settings['id']] = Board(settings)
-
-        for board in BOARDS_ALL:  # докидываем скрытых борд, на которые Абу не дает настроек
-            if board not in self._boards.keys():
-                self._boards[board] = Board({'id': board})
-
-        return True
 
     def get_board(self, board=None):
         """
@@ -313,7 +48,8 @@ class DvachApi:
         if not (board and self.board_exist(board)):  # pragma: no cover
             board = self.board.id
 
-        threads = self.__Session.get(board, 'threads.json').threads
+        threads = self.__ApiClient.request(
+            url=f'{CHAN_URL}/{board}/threads.json').threads
 
         return [Thread(thread) for thread in threads]
 
@@ -330,7 +66,8 @@ class DvachApi:
         if not (board and self.board_exist(board)):  # pragma: no cover
             board = self.board.id
 
-        posts = self.__Session.get(board, f'res/{thread}.json').threads
+        posts = self.__ApiClient.request(
+            url=f'{CHAN_URL}/{board}/res/{thread}.json').threads
 
         return [Post(post) for post in posts[0].posts]
 
@@ -345,14 +82,18 @@ class DvachApi:
         if not (board and self.board_exist(board)):  # pragma: no cover
             board = self.board.id
 
-        threads = self.__Session.get(board, 'threads.json').threads
+        threads = self.__ApiClient.request(
+            url=f'{CHAN_URL}/{board}/threads.json').threads
 
         if method == 'views':
-            threads = sorted(threads, key=lambda thread: (thread['views'], thread['score']), reverse=True)
+            threads = sorted(threads, key=lambda thread: (
+                thread['views'], thread['score']), reverse=True)
         elif method == 'score':
-            threads = sorted(threads, key=lambda thread: (thread['score'], thread['views']), reverse=True)
+            threads = sorted(threads, key=lambda thread: (
+                thread['score'], thread['views']), reverse=True)
         elif method == 'posts':
-            threads = sorted(threads, key=lambda thread: (thread['posts_count'], thread['views']), reverse=True)
+            threads = sorted(threads, key=lambda thread: (
+                thread['posts_count'], thread['views']), reverse=True)
         else:
             return []
 
@@ -363,12 +104,15 @@ class DvachApi:
         Авторизация пасскода
         :param usercode: Пасскод
         """
-        url = url_join(URL, 'makaba/makaba.fcgi')
+        url = f'{CHAN_URL}/makaba/makaba.fcgi'
         payload = {
             'task': 'auth',
             'usercode': usercode
         }
-        response = self.__Session.post(url=url, data=payload)
+        response = self.__ApiClient.request(
+            method='post', url=url,
+            data=payload)
+
         self.passcode_data = response.cookies['usercode_nocaptcha']
 
         return True
@@ -383,6 +127,15 @@ class DvachApi:
         """
         if passcode:
             message.payload.usercode = self.passcode_data
+            # при наличии файлов- проверяем их
+            if message.files != {'': ''}:
+                if len(message.files) > 8:
+                    raise ex.ExtraFilesError(files_len=len(message.files),
+                                             passcode=True)
+                elif message.filesize.size > 60:
+                    raise ex.FileSizeError(files_size=message.filesize.size,
+                                           passcode=True)
+
         elif captcha:
             captcha_payload = {
                 'captcha_type': captcha.captcha_type,
@@ -390,35 +143,45 @@ class DvachApi:
                 '2chaptcha_value': captcha.captcha_value
             }
             message.payload.update(captcha_payload)
-            if message.files != {'': ''}:  # TODO: Убираем лишние файлы, если капча или выбрасывать исключение?
-                while len(message.files) > 4:
-                    message.files.popitem()
+
+            # при наличии файлов проверяем их
+            if message.files != {'': ''}:
+                if len(message.files) > 4:
+                    raise ex.ExtraFilesError(files_len=len(message.files),
+                                             passcode=False)
+                elif message.filesize.size > 20:
+                    raise ex.FileSizeError(files_size=message.filesize.size,
+                                           passcode=False)
+
         else:
-            return False
+            raise ex.AuthRequiredError()
 
         try:
-            response = self.__Session.post(url='makaba/posting.fcgi',
-                                           data=message.payload,
-                                           files=message.files)
+            response = self.__ApiClient.request(
+                method='post',
+                url=f'{CHAN_URL}/makaba/posting.fcgi',
+                data=message.payload,
+                files=message.files)
         except Exception as e:
             print('Error send post: {msg}'.format(msg=e))
             return False
         else:
             return response
 
-    def find(self, board=None, thread=None, patterns=None, antipatterns=None):
+    def find(self, board=None, thread=None, patterns=None, anti_patterns=None):
         """
         Поиск тредов по заданным строкам в шапке
-        :param board: ИД борды на которой нужно искать тред по заданным строкам в ОП-посте
-        :param thread: ИД треда или объект типа Thread в постах которого нужно провести поиск
+        :param board: ИД борды на которой искать тред по словам в ОП-посте
+        :param thread: ИД треда или объект типа Thread для поиска
+
         :param patterns: Список фраз для поиска в шапке
-        :param antipatterns: Список фраз которые не должны всречаться в шапке
+        :param anti_patterns: Список фраз которые не должны всречаться в шапке
         :return: Список тредов удовлетворяющих условиям
         """
         if patterns is None:
             patterns = []
-        if antipatterns is None:
-            antipatterns = []
+        if anti_patterns is None:
+            anti_patterns = []
 
         if not (board and self.board_exist(board)):  # pragma: no cover
             board = self.board.id
@@ -426,15 +189,25 @@ class DvachApi:
         if isinstance(thread, Thread):
             thread = thread.num
 
+        # TODO: Тут покрасивее сделать надо
         if thread:
             posts = self.get_thread(board=board, thread=thread)
-            matched = [post for post in posts if any(subs in post.comment.lower() for subs in patterns) and all(
-                subs not in post.comment.lower for subs in antipatterns)]
+            matched = [
+                post for post in posts if
+                any(subs in post.comment.lower() for subs in patterns) and
+                all(subs not in post.comment.lower for subs in anti_patterns)
+            ]
         else:
             threads = self.get_board(board=board)
-            matched = [thread for thread in threads if
-                       any(subs in thread.post.comment.lower() for subs in patterns) and all(
-                           subs not in thread.post.comment.lower() for subs in antipatterns)]
+            matched = [
+                thread for thread in threads if
+                any(subs in thread.post.comment.lower() for subs in
+                    patterns) and
+                all(
+                    subs not in thread.post.comment.lower() for subs in
+                    anti_patterns
+                )
+            ]
 
         return matched
 
@@ -443,7 +216,7 @@ class DvachApi:
         Установка заголовка запросов на лету
         :param headers:
         """
-        self.__Session.update_headers(headers)
+        self.__ApiClient.update_headers(headers)
         return True
 
     def set_proxies(self, proxies=None):
@@ -451,17 +224,16 @@ class DvachApi:
         Установка прокси на лету
         :param proxies:
         """
-        self.__Session.update_proxies(proxies)
+        self.__ApiClient.update_proxies(proxies)
         return True
 
-    @staticmethod
-    def board_exist(board):
+    def board_exist(self, board):
         """
         Проверка существование доски в списке всех досок
         :param board: ИД доски. Например 'b'
         :return: boolean
         """
-        return board in BOARDS_ALL
+        return board in self.boards.keys()
 
     def __repr__(self):
-        return '<Api: {board}>'.format(board=self.board.id)
+        return f'<Api: {self.board.id}>'
