@@ -1,4 +1,9 @@
-from .api import *
+from .api import DvachApi
+from .captcha import CaptchaHelper
+from .client import ApiClient
+from .models import Message
+from .settings import CHAN_URL
 
 __author__ = 'anonim-legivon'
-__version__ = '0.9.0'
+
+__version__ = '1.0.0'
