@@ -1,7 +1,7 @@
 """2ch.hk API Utils"""
 
 
-def listmerge(boards_dict):
+def list_merge(boards_dict):
     """
     Merge dict of boards into list
     :param boards_dict: dict of boards
